clova-invoice-scanner/clova-service/requirements.txt
# FastAPI and server
fastapi==0.100.0
uvicorn[standard]==0.23.0
python-multipart==0.0.6
python-dotenv==1.0.0

# AI and ML libraries
torch>=1.13.0
torchvision>=0.14.0
transformers>=4.25.0
donut-python>=1.0.9
craft-text-detector>=0.4.3

# Image processing
opencv-python>=3.4.8.29,<4.5.4.62
Pillow>=9.0.0
numpy>=1.21.0
scikit-image>=0.19.0

# HTTP client
httpx>=0.24.0
aiofiles>=23.0.0

# Utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Logging and monitoring
structlog>=23.0.0
prometheus-client>=0.17.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# Optional: GPU support
# torch>=1.13.0+cu117 -f https://download.pytorch.org/whl/torch_stable.html
# torchvision>=0.14.0+cu117 -f https://download.pytorch.org/whl/torch_stable.html 